\name{data_gaussian}
\alias{data_gaussian}
\docType{data}
\title{
Gaussian data. 
}
\description{
Sample of size \eqn{n=100} from a multivariate gaussian distribution with \eqn{p=50} variables.
}
\usage{data("data_gaussian")}
\format{
  The format is:
 num [1:50, 1:100] 1.001 -0.21 0.513 0.166 2.135 ...
}

\examples{
data(data_gaussian)
}
\keyword{datasets}
