\name{data_multinomial}
\alias{data_multinomial}
\docType{data}
\title{
Multinomial data.
}
\description{
Sample of size \eqn{n=100} from a multinomial distribution with \eqn{p=100} variables.
}
\usage{data("data_multinomial")}
\format{
  The format is:
 int [1:100, 1:100] 8 10 5 3 2 8 3 5 8 3 ...
}

\examples{
data(data_multinomial)
}
\keyword{datasets}
